docs(skeleton): Migrate Skeleton MDX from canvas-site
