+++ modules/react/skeleton/stories/Skeleton.stories.mdx
@@ -1,0 +1,125 @@
+import {Meta} from '@storybook/addon-docs/blocks';
+
+import {Skeleton} from '@workday/canvas-kit-react/skeleton';
+
+import {Basic} from './examples/Basic';
+import {Color} from './examples/Color';
+import {Header} from './examples/Header';
+import {NonSkeletonChildren} from './examples/NonSkeletonChildren';
+import {Shape} from './examples/Shape';
+import {Text} from './examples/Text';
+
+<Meta title="Components/Indicators/Skeleton/React" component={Skeleton} />
+
+# Canvas Kit Skeleton
+
+A Skeleton is a low-fidelity visual placeholder that represents the loading of interface elements
+before they have displayed on the page. Appearing as a blank version of the page, it mitigates focus
+on the loading process and improves the user’s perceived load time.
+
+[> Workday Design Reference](https://design.workday.com/components/indicators/skeleton-loader)
+
+## Installation
+
+```sh
+yarn add @workday/canvas-kit-react
+```
+
+## Usage
+
+### Basic Example
+
+Skeleton includes a container component (`Skeleton`) as well as a number of subcomponents
+(`Skeleton.Header`, `Skeleton.Text`, and `Skeleton.Shape`) which can be used as placeholders for
+different types of content.
+
+Here's an example of how Skeleton components can be combined to create a loading placeholder for
+Card content. Press the **Simulate Loading** button to simulate the loading of the content
+(customize the loading time using the **Load Time** field), or check the **Loading** check box to
+force the Skeleton to display.
+
+<ExampleCodeBlock code={Basic} />
+
+### Color
+
+All Skeleton subcomponents accept a `backgroundColor` prop which can be used to specify the color of
+the subcomponent.
+
+<ExampleCodeBlock code={Color} />
+
+### Non-Skeleton Children
+
+Note that non-Skeleton components can be included as children of the `Skeleton` container component
+if you'd like to customize your Skeleton further.
+
+<ExampleCodeBlock code={NonSkeletonChildren} />
+
+## Components
+
+### Skeleton
+
+#### Usage
+
+Skeleton subcomponents **must** be wrapped by the `Skeleton` container component.
+
+```tsx
+<Skeleton>
+  <Skeleton.Header />
+  <Skeleton.Text />
+</Skeleton>
+```
+
+`Skeleton` places its children in a container element with `aria-hidden` set to `true` and announces
+itself using a visually hidden element.
+
+#### Props
+
+Undocumented props are spread to the outermost element rendered by `Skeleton`.
+
+<ArgsTable of={Skeleton} />
+
+### Skeleton.Header
+
+#### Usage
+
+`Skeleton.Header` renders a placeholder for header content such as headings.
+
+<ExampleCodeBlock code={Header} />
+
+#### Props
+
+Undocumented props are spread to the outermost element rendered by `Skeleton.Header`.
+
+<ArgsTable of={Skeleton.Header} />
+
+### Skeleton.Text
+
+#### Usage
+
+`Skeleton.Text` renders a placeholder for text content such as paragraphs. Each placeholder line has
+a width of `100%` and a fixed height of `21px`, with the last line having a width of `60%` if there
+are multiple lines.
+
+<ExampleCodeBlock code={Text} />
+
+#### Props
+
+Undocumented props are spread to the outermost element rendered by `Skeleton.Text`.
+
+<ArgsTable of={Skeleton.Text} />
+
+### Skeleton.Shape
+
+#### Usage
+
+`Skeleton.Shape` renders a placeholder for content which can be represented as a shape such as a
+circular icon. Set the `height`, `width`, and `borderRadius` props of the `Skeleton.Shape` to create
+various rectangular and circular shapes.
+
+<ExampleCodeBlock code={Shape} />
+
+#### Props
+
+Undocumented props are spread to the outermost element rendered by `Skeleton.Shape`.
+
+<ArgsTable of={Skeleton.Shape} />
