+++ modules/react/skeleton/stories/examples/Header.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import {Skeleton} from '@workday/canvas-kit-react/skeleton';
+
+export const Header = () => {
+  return (
+    <Skeleton>
+      <Skeleton.Header />
+    </Skeleton>
+  );
+};
