+++ modules/react/skeleton/stories/examples/Text.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import {Skeleton} from '@workday/canvas-kit-react/skeleton';
+
+export const Text = () => {
+  return (
+    <Skeleton>
+      <Skeleton.Text />
+    </Skeleton>
+  );
+};
